Fixed a small bug that caused windows to sometimes disappear on startup.
Windawesome now uses a global hotkey that it activates when it needs to activate an application, as Windows gives foreground activation permission when it sends WM_HOTKEY to an application. This should work better, faster and, most importantly, shouldn't hang the UI when switching to a hung application, which wasn't the case when using AttachThreadInput.

Added "sealed" to some of the classes. Don't know why.